refactor: move useDebounce to hooks dir
